requirements.txt
# FastAPI and ASGI server
fastapi>=0.104.1,<0.105.0
uvicorn[standard]>=0.24.0,<0.25.0

# MongoDB async driver
motor>=3.3.2,<3.4.0
pymongo[srv]>=4.6.0,<5.0.0

# Google Gemini API
google-generativeai>=0.3.2

# Environment and configuration management
python-dotenv>=1.0.0,<2.0.0
pydantic[email]>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0

# Web scraping dependencies
aiohttp[speedups]>=3.9.0,<4.0.0
beautifulsoup4>=4.12.0,<5.0.0
aiofiles>=23.2.0,<24.0.0
robotparser>=1.0.0

# System monitoring
psutil>=5.9.0,<6.0.0

# Language detection (optional)
langdetect>=1.0.9,<2.0.0

# NLP and text processing (optional)
textstat>=0.7.3,<0.8.0

# Development dependencies (optional)
pytest>=7.4.3,<8.0.0
pytest-asyncio>=0.21.1,<0.22.0
aioresponses>=0.7.4,<1.0.0
mongomock>=4.1.0,<5.0.0
black>=23.11.0,<24.0.0
isort>=5.12.0,<6.0.0
flake8>=6.1.0,<7.0.0
